Added Secondary Index for yarch tables
- only one index per table supported for the moment
- simplified query execution: the where conditions are always verified and not relied on the index
- removed the complicated strictStart/strictEnd from range iterators
- replaced the DataOutputStream with ByteArray in table stream serialization
- renamed decodeShort into decodeUnsignedShort
- in V3 tables use null terminated UTF string encoding rather than prefix based such that strings can be used as primary and secondary keys

+++ yamcs-core/src/test/java/org/yamcs/yarch/StreamSelectBinaryFunctionTest.java
@@ -59,7 +59,7 @@ public class StreamSelectBinaryFunctionTest extends YarchTestCase {
         assertEquals(n, l.size());
         for (int i = 0; i < n; i++) {
             Tuple t = l.get(i);
-            int x = (Integer) t.getColumn(0);
+            int x = (Short) t.getColumn(0);
             assertEquals(i, x);
         }
     }
